requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-functions==1.20.0
fastapi==0.112.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
pymssql==2.3.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
